+++ lib/postcss-mixins/clearfix.js
@@ -1,0 +1,7 @@
+module.exports = {
+  '&::after': {
+    content: '""',
+    display: 'table',
+    clear: 'both'
+  }
+}
